GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - src/views/HomeView.vue

Your previous branch was: refs/heads/main

The sha for that commit was: 1db755030594ba0f95a7700ec6962f68ef48b91e

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch

feat: Update WelcomeItem content

- Updated the Tooling section in the WelcomeItem component to include more detailed information about the project's setup and recommended tools.
- Added a line break after the Tooling section for better readability.
- Revised the content in the Ecosystem section of the WelcomeItem component to include a list of official tools and libraries for the project.
- Improved the Community section in the WelcomeItem component by providing more detailed information and resources for users who may need help or support.
refactor: Remove unnecessary styling in TheWelcome.vue

The changes remove the unnecessary styling for the body element in TheWelcome.vue file. The styling was not affecting any visible element in the component and was therefore removed to improve code cleanliness and maintainability.
style: Change background color to black

Changed the background color of the body to black in the style section to improve the overall aesthetic and user experience.